refactor: Extract Spaced.jsx
