+++ frontend/src/components/Spaced.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Spaced = ({elements}) => (
+  elements
+    .map((x, index) => <span key={index}>{x}</span>)
+    .reduce((prev, curr) => [
+      prev,
+      <span key = {prev+curr} className = 'spacer-dot' />,
+      curr
+    ])
+);
+
+export default Spaced;
